Split into separate test suite
