+++ core/test/variable_test.cpp
@@ -8,7 +8,6 @@
 #include "scipp/core/dimensions.h"
 #include "scipp/core/except.h"
 #include "scipp/core/variable.h"
-#include "scipp/core/variable.tcc"
 
 using namespace scipp;
 using namespace scipp::core;
@@ -89,29 +88,6 @@ TEST(Variable, span_references_Variable) {
   EXPECT_EQ(span.size(), 2);
   span[0] = 1.0;
   EXPECT_EQ(observer[0], 1.0);
-}
-
-struct CustomType {
-public:
-  CustomType() {}
-  CustomType(const int &) {}
-  bool operator==(const CustomType &) const { return true; }
-};
-
-// Instantiate Variable type. Test template instantiation with custom type
-// argument.
-INSTANTIATE_VARIABLE(CustomType)
-// Instantiate VariableConstProxy type. Test template instantiation with
-// custom type argument.
-INSTANTIATE_SLICEVIEW(CustomType)
-
-TEST(Variable, use_custom_templates) {
-  auto input_values = std::initializer_list<CustomType>{1, 2};
-  auto var = Variable{{Dim::X, 2}, input_values};
-  // Check for bad cast or other built-in implicit type assumptions
-  EXPECT_NO_THROW(var.values<CustomType>());
-  VariableConstProxy slice = var.slice(Slice(Dim::X, 0));
-  EXPECT_NO_THROW(slice.values<CustomType>());
 }
 
 class Variable_comparison_operators : public ::testing::Test {
